feat: implement create aws security group by boto3
this design the appliction action define on .env

we try using the '.env.bak' mocking the '.env' file, just want to the presentation with '.env' need format

+++ lab/ec2.py
@@ -2,14 +2,17 @@ import boto3
 import logging
 import logging.config
 from dotenv import dotenv_values
+from botocore.exceptions import ClientError
 
-# initialize logging module
-env = dotenv_values(".env")
+# Note: initialize logging module
+__env__ = dotenv_values(".env")
+__author_name__ = __env__.get("author")
+__logger_tag__=__env__.get('logger_tag')
 
-logger_tag=env.get('logger_tag')
 logging.config.fileConfig(fname='log.conf')
-logger = logging.getLogger(f'{logger_tag}')
+logger= logging.getLogger(f'{__logger_tag__}')
 
+# Note: initialize creating fuctuon
 def create_security_group(input_ec2: boto3, input_sg_name: str, input_sg_descrbe: str, input_vpc_id: str):
     """Create the AWS Security Group by Boto3 EC2.
 
@@ -19,31 +22,36 @@ def create_security_group(input_ec2: boto3, input_sg_name: str, input_sg_descrbe
         input_sg_descrbe (str): _description_
         input_vpc_id (str): _description_
     """    
-
     try:
-        # security_group_params = {
-        #     'GroupName': input_sg_name,     
-        #     'Description': input_sg_descrbe, 
-        #     'VpcId': input_vpc_id,            
-        # }
-        # response = input_ec2.create_security_group(**security_group_params)
-        # security_group_id = response['GroupId']
+        response_checking = input_ec2.describe_security_groups(Filters=[
+            {'Name': f'tag:Name', 'Values': [input_sg_name]}
+            ])
+        
+        if response_checking['SecurityGroups']:
+            security_group_params = {
+                'GroupName': input_sg_name,     
+                'Description': input_sg_descrbe, 
+                'VpcId': input_vpc_id,            
+            }
+            response_createing = input_ec2.create_security_group(**security_group_params)
+            security_group_id = response_createing['GroupId']
+            tag_params = {
+                'Resources': [security_group_id],  
+                'Tags': [
+                    {'Key': 'Name', 'Value': input_sg_name},
+                    {'Key': 'Author', 'Value': __author_name__}
+                ]
+            }
+            ec2.create_tags(**tag_params)
+            logger.info(f"Security Group Name: '{input_sg_name}({security_group_id})' Created!")
+        else:
+            logger.info(f"The security group '{input_sg_name}' already exists for VPC '{input_vpc_id}'")
 
-        # tag_params = {
-        #     'Resources': [security_group_id],  
-        #     'Tags': [
-        #         {'Key': 'Name', 'Value': input_sg_name} 
-        #     ]
-        # }
-        # ec2.create_tags(**tag_params)
-
-        # Mock Security Group ID
-        output_sg_id = 'mock_security_group_id'
-        logger.info(f"Security Group Name: '{input_sg_name}({output_sg_id})' Created!")
     except Exception as err:
         logger.info(f"Security Group Name: '{input_sg_name}' Create Failed!")
         logger.info(f"Unexpected: {err}, {type(err)}")    
 
+# Note: initialize deleting fuctuon
 def delete_security_group(input_ec2: boto3, input_sg_name: str):
     """Delete the Security Group by Boto3 EC2.
 
@@ -53,33 +61,34 @@ def delete_security_group(input_ec2: boto3, input_sg_name: str):
     """   
 
     try:
-        # response = input_ec2.describe_security_groups(Filters=[
-        #     {'Name': f'tag:Name', 'Values': [input_sg_name]}
-        #     ])
-        # if response['SecurityGroups']:
-        #     security_group = response['SecurityGroups'][0]
-        #     security_group_id = security_group['GroupId']
-        #     input_ec2.delete_security_group(GroupId=security_group_id)
-        #     logger.info(f"Security Group Name: '{input_sg_name}({output_sg_id})' Deleted!")
-        # else:
-        #     logger.info(f"No security groups found with the tag '{'Name'}={input_sg_name}'.")
+        response_checking = input_ec2.describe_security_groups(Filters=[
+            {'Name': f'tag:Name', 'Values': [input_sg_name]}
+            ])
+        if response_checking['SecurityGroups']:
+            security_group = response_checking['SecurityGroups'][0]
+            security_group_id = security_group['GroupId']
+            input_ec2.delete_security_group(GroupId=security_group_id)
+            logger.info(f"Security Group Name: '{input_sg_name}({security_group_id})' Deleted!")
+        else:
+            logger.info(f"No security groups found with the tag '{'Name'}={input_sg_name}'.")
 
-        # Mock Security Group ID
-        output_sg_id = 'mock_security_group_id'    
-        logger.info(f"Security Group Name: '{input_sg_name}({output_sg_id})' Deleted!")
     except Exception as err:
         logger.info(f"Security Group Name: '{input_sg_name}' Delete Failed!")
         logger.info(f"Unexpected: {err}, {type(err)}")
     
-
-vpc_id = env.get('vpc_id')
-sg_name = env.get('sg_name')
-sg_describe = env.get('sg_describe')
-ec2 = 'mock boto3 client object'
-# ec2 = boto3.client('ec2')
+# Note: the application start
+vpc_id = __env__.get('vpc_id')
+sg_name = __env__.get('sg_name')
+sg_describe = __env__.get('sg_describe')
+action = __env__.get('appliction_action')
+ec2 = boto3.client('ec2')
 
 try:
-    create_security_group(ec2, sg_name, sg_describe, vpc_id)
-    delete_security_group(ec2, sg_name)
+    if (action=='create'):
+        create_security_group(ec2, sg_name, sg_describe, vpc_id)
+    elif (action=='delete'):
+        delete_security_group(ec2, sg_name)
+    else:
+        logger.info(f"The action '{action}' is not supported.")
 except Exception as err:
     logger.info(f"Unexpected: {err}, {type(err)}")
